+++ x/foundation/msgs_test.go
@@ -1,6 +1,7 @@
 package foundation_test
 
 import (
+	"fmt"
 	"testing"
 	"time"
 
@@ -9,7 +10,9 @@ import (
 	"github.com/line/lbm-sdk/crypto/keys/secp256k1"
 	"github.com/line/lbm-sdk/testutil/testdata"
 	sdk "github.com/line/lbm-sdk/types"
+	"github.com/line/lbm-sdk/x/auth/legacy/legacytx"
 	"github.com/line/lbm-sdk/x/foundation"
+	"github.com/line/lbm-sdk/x/stakingplus"
 )
 
 func TestMsgUpdateParams(t *testing.T) {
@@ -682,5 +685,235 @@ func TestMsgGovMint(t *testing.T) {
 		require.NoError(t, err, name)
 
 		require.Equal(t, []sdk.AccAddress{tc.authority}, msg.GetSigners(), name)
+	}
+}
+
+func TestAminoJSON(t *testing.T) {
+	addrs := make([]sdk.AccAddress, 3)
+	for i := range addrs {
+		addrs[i] = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	}
+
+	testCases := map[string]struct {
+		msg      legacytx.LegacyMsg
+		expected string
+	}{
+		"MsgFundTreasury": {
+			&foundation.MsgFundTreasury{
+				From:   addrs[0].String(),
+				Amount: sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.OneInt())),
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgFundTreasury\",\"value\":{\"amount\":[{\"amount\":\"1\",\"denom\":\"stake\"}],\"from\":\"%s\"}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String()),
+		},
+		"MsgVote": {
+			&foundation.MsgVote{
+				ProposalId: 1,
+				Voter:      addrs[0].String(),
+				Option:     foundation.VOTE_OPTION_YES,
+				Metadata:   "I'm YES",
+				Exec:       foundation.Exec_EXEC_UNSPECIFIED,
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgVote\",\"value\":{\"metadata\":\"I'm YES\",\"option\":1,\"proposal_id\":\"1\",\"voter\":\"%s\"}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String()),
+		},
+		"MsgExec": {
+			&foundation.MsgExec{
+				ProposalId: 1,
+				Signer:     addrs[0].String(),
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgExec\",\"value\":{\"proposal_id\":\"1\",\"signer\":\"%s\"}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String()),
+		},
+		"MsgLeaveFoundation": {
+			&foundation.MsgLeaveFoundation{Address: addrs[0].String()},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgLeaveFoundation\",\"value\":{\"address\":\"%s\"}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String()),
+		},
+		"MsgWithdrawProposal": {
+			&foundation.MsgWithdrawProposal{
+				ProposalId: 1,
+				Address:    addrs[0].String(),
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgWithdrawProposal\",\"value\":{\"address\":\"%s\",\"proposal_id\":\"1\"}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String()),
+		},
+	}
+
+	for name, tc := range testCases {
+		tc := tc
+
+		t.Run(name, func(t *testing.T) {
+			require.Equal(t, tc.expected, string(legacytx.StdSignBytes("foo", 1, 1, 1, legacytx.StdFee{}, []sdk.Msg{tc.msg}, "memo")))
+		})
+	}
+}
+
+func TestMsgSubmitProposalAminoJSON(t *testing.T) {
+	addrs := make([]sdk.AccAddress, 2)
+	for i := range addrs {
+		addrs[i] = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	}
+
+	var proposer = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	testCases := map[string]struct {
+		msg      sdk.Msg
+		expected string
+	}{
+		"MsgUpdateParams": {
+			&foundation.MsgUpdateParams{
+				Authority: addrs[0].String(),
+				Params:    foundation.Params{FoundationTax: sdk.ZeroDec()},
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgUpdateParams\",\"value\":{\"authority\":\"%s\",\"params\":{\"foundation_tax\":\"0.000000000000000000\"}}}],\"metadata\":\"MsgUpdateParams\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String(), proposer.String()),
+		},
+		"MsgWithdrawFromTreasury": {
+			&foundation.MsgWithdrawFromTreasury{
+				Authority: addrs[0].String(),
+				To:        addrs[1].String(),
+				Amount:    sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(1000000))),
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgWithdrawFromTreasury\",\"value\":{\"amount\":[{\"amount\":\"1000000\",\"denom\":\"stake\"}],\"authority\":\"%s\",\"to\":\"%s\"}}],\"metadata\":\"MsgWithdrawFromTreasury\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String(), addrs[1].String(), proposer.String()),
+		},
+		"MsgUpdateMembers": {
+			&foundation.MsgUpdateMembers{
+				Authority: addrs[0].String(),
+				MemberUpdates: []foundation.MemberRequest{{
+					Address: addrs[1].String(),
+				}},
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgUpdateMembers\",\"value\":{\"authority\":\"%s\",\"member_updates\":[{\"address\":\"%s\"}]}}],\"metadata\":\"MsgUpdateMembers\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String(), addrs[1].String(), proposer.String()),
+		},
+		"MsgRevoke": {
+			&foundation.MsgRevoke{
+				Authority:  addrs[0].String(),
+				Grantee:    addrs[1].String(),
+				MsgTypeUrl: foundation.ReceiveFromTreasuryAuthorization{}.MsgTypeURL(),
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgRevoke\",\"value\":{\"authority\":\"%s\",\"grantee\":\"%s\",\"msg_type_url\":\"/lbm.foundation.v1.MsgWithdrawFromTreasury\"}}],\"metadata\":\"MsgRevoke\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String(), addrs[1].String(), proposer.String()),
+		},
+		"MsgGovMint": {
+			&foundation.MsgGovMint{
+				Authority: addrs[0].String(),
+				Amount:    sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10))),
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgGovMint\",\"value\":{\"amount\":[{\"amount\":\"10\",\"denom\":\"stake\"}],\"authority\":\"%s\"}}],\"metadata\":\"MsgGovMint\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String(), proposer.String()),
+		},
+	}
+
+	for name, tc := range testCases {
+		tc := tc
+
+		t.Run(name, func(t *testing.T) {
+			proposalMsg := &foundation.MsgSubmitProposal{
+				Proposers: []string{proposer.String()},
+				Metadata:  name,
+				Exec:      foundation.Exec_EXEC_TRY,
+			}
+			err := proposalMsg.SetMsgs([]sdk.Msg{tc.msg})
+			require.NoError(t, err)
+			require.Equal(t, tc.expected, string(legacytx.StdSignBytes("foo", 1, 1, 1, legacytx.StdFee{}, []sdk.Msg{proposalMsg}, "memo")))
+		})
+	}
+}
+
+func TestMsgUpdateDecisionPolicyAminoJson(t *testing.T) {
+	var (
+		authority = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+		proposer  = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	)
+
+	testCases := map[string]struct {
+		policy   foundation.DecisionPolicy
+		expected string
+	}{
+		"ThresholdDecisionPolicy": {
+			&foundation.ThresholdDecisionPolicy{
+				Threshold: sdk.OneDec(),
+				Windows: &foundation.DecisionPolicyWindows{
+					VotingPeriod: time.Hour,
+				},
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgUpdateDecisionPolicy\",\"value\":{\"authority\":\"%s\",\"decision_policy\":{\"type\":\"lbm-sdk/ThresholdDecisionPolicy\",\"value\":{\"threshold\":\"1.000000000000000000\",\"windows\":{\"min_execution_period\":\"0\",\"voting_period\":\"3600000000000\"}}}}}],\"metadata\":\"ThresholdDecisionPolicy\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", authority, proposer),
+		},
+		"PercentageDecisionPolicy": {
+			&foundation.PercentageDecisionPolicy{
+				Percentage: sdk.OneDec(),
+				Windows: &foundation.DecisionPolicyWindows{
+					VotingPeriod: time.Hour,
+				},
+			},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgUpdateDecisionPolicy\",\"value\":{\"authority\":\"%s\",\"decision_policy\":{\"type\":\"lbm-sdk/PercentageDecisionPolicy\",\"value\":{\"percentage\":\"1.000000000000000000\",\"windows\":{\"min_execution_period\":\"0\",\"voting_period\":\"3600000000000\"}}}}}],\"metadata\":\"PercentageDecisionPolicy\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", authority, proposer),
+		},
+	}
+
+	for name, tc := range testCases {
+		tc := tc
+
+		t.Run(name, func(t *testing.T) {
+			policyMsg := &foundation.MsgUpdateDecisionPolicy{
+				Authority: authority.String(),
+			}
+			err := policyMsg.SetDecisionPolicy(tc.policy)
+			require.NoError(t, err)
+
+			err = policyMsg.ValidateBasic()
+			require.NoError(t, err)
+
+			proposalMsg := &foundation.MsgSubmitProposal{
+				Proposers: []string{proposer.String()},
+				Metadata:  name,
+				Exec:      foundation.Exec_EXEC_TRY,
+			}
+			err = proposalMsg.SetMsgs([]sdk.Msg{policyMsg})
+			require.NoError(t, err)
+
+			require.Equal(t, tc.expected, string(legacytx.StdSignBytes("foo", 1, 1, 1, legacytx.StdFee{}, []sdk.Msg{proposalMsg}, "memo")))
+		})
+	}
+}
+
+func TestMsgGrantAminoJson(t *testing.T) {
+	var (
+		operator      = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+		grantee       = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+		proposer      = sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+		validatorAddr = sdk.ValAddress(secp256k1.GenPrivKey().PubKey().Address())
+	)
+
+	testCases := map[string]struct {
+		authorization foundation.Authorization
+		expected      string
+	}{
+		"ReceiveFromTreasuryAuthorization": {
+			&foundation.ReceiveFromTreasuryAuthorization{},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgGrant\",\"value\":{\"authority\":\"%s\",\"authorization\":{\"type\":\"lbm-sdk/ReceiveFromTreasuryAuthorization\",\"value\":{}},\"grantee\":\"%s\"}}],\"metadata\":\"ReceiveFromTreasuryAuthorization\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", operator.String(), grantee.String(), proposer.String()),
+		},
+		"CreateValidatorAuthorization": {
+			&stakingplus.CreateValidatorAuthorization{ValidatorAddress: validatorAddr.String()},
+			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgGrant\",\"value\":{\"authority\":\"%s\",\"authorization\":{\"type\":\"lbm-sdk/CreateValidatorAuthorization\",\"value\":{\"validator_address\":\"%s\"}},\"grantee\":\"%s\"}}],\"metadata\":\"CreateValidatorAuthorization\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", operator.String(), validatorAddr.String(), grantee.String(), proposer.String()),
+		},
+	}
+
+	for name, tc := range testCases {
+		tc := tc
+
+		t.Run(name, func(t *testing.T) {
+			grantMsg := &foundation.MsgGrant{
+				Authority: operator.String(),
+				Grantee:   grantee.String(),
+			}
+			err := grantMsg.SetAuthorization(tc.authorization)
+			require.NoError(t, err)
+
+			err = grantMsg.ValidateBasic()
+			require.NoError(t, err)
+
+			proposalMsg := &foundation.MsgSubmitProposal{
+				Proposers: []string{proposer.String()},
+				Metadata:  name,
+				Exec:      foundation.Exec_EXEC_TRY,
+			}
+			err = proposalMsg.SetMsgs([]sdk.Msg{grantMsg})
+			require.NoError(t, err)
+
+			require.Equal(t, tc.expected, string(legacytx.StdSignBytes("foo", 1, 1, 1, legacytx.StdFee{}, []sdk.Msg{proposalMsg}, "memo")))
+		})
 	}
 }